Refactor to separate comparison wrapper functions
