+++ application/common/errors.py
@@ -1,2 +1,6 @@
 class ValidationException(Exception):
     pass
+
+
+class DynamodbException(Exception):
+    pass
